Refactor imports and DI modules
This commit introduces the following changes:
- Removes unused imports across various files.
- Deletes `shared_pref_di.kt` from `core/sharedpreference` and recreates it in `core/di`.
- Moves `ISharedPreferences.kt` and `SharedPreferencesImpl.kt` from `core/sharedpreference` to `core/local/sharedpreference`.
- Adds `FirebaseAuth` provider to `FirebaseModule.kt`.
- Removes `FirebaseModule` object from `AuthenticationDi.kt` as it's now provided globally.
- Updates `deploymentTargetSelector.xml` by removing specific run configuration states.
- Removes `DeleteAllScrapsUseCase` from `AddItemModule.kt`.
