+++ app/src/main/java/com/delighted2wins/souqelkhorda/core/components/DirectionalText.kt
@@ -2,7 +2,8 @@ package com.delighted2wins.souqelkhorda.core.components
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.material3.*
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.ui.Modifier
