+++ app/src/main/java/com/delighted2wins/souqelkhorda/core/di/FirebaseModule.kt
@@ -1,14 +1,20 @@
 package com.delighted2wins.souqelkhorda.core.di
 
+import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.FirebaseFirestore
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 object FirebaseModule {
+
+    @Provides
+    @Singleton
+    fun provideFirebaseAuth(): FirebaseAuth = FirebaseAuth.getInstance()
 
     @Provides
     fun provideFirestore(): FirebaseFirestore = FirebaseFirestore.getInstance()
