+++ app/src/main/java/com/delighted2wins/souqelkhorda/core/di/shared_pref_di.kt
@@ -1,5 +1,4 @@
-package com.delighted2wins.souqelkhorda.core.sharedpreference
-
+package com.delighted2wins.souqelkhorda.core.di
 
 import com.abdok.atmosphere.data.local.sharedPreference.ISharedPreferences
 import com.example.anees.data.local.sharedpreference.SharedPreferencesImpl
@@ -9,11 +8,9 @@ import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
 import javax.inject.Singleton
 
-
 @Module
 @InstallIn(SingletonComponent::class)
-abstract class shared_pref_di {
-
+abstract class shared_pref_di{
     @Binds
     @Singleton
     abstract fun bindISharedPreferences(
