+++ app/src/main/java/com/delighted2wins/souqelkhorda/core/local/db/ScrapDao.kt
@@ -1,7 +1,6 @@
 package com.delighted2wins.souqelkhorda.core.local.db
 
 import androidx.room.Dao
-import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Query
 import com.delighted2wins.souqelkhorda.features.sale.domain.entities.Scrap
