+++ app/src/main/java/com/delighted2wins/souqelkhorda/features/market/data/remote/MarketRemoteDataSource.kt
@@ -2,8 +2,8 @@ package com.delighted2wins.souqelkhorda.features.market.data.remote
 
 import com.delighted2wins.souqelkhorda.features.market.domain.entities.ScrapOrder
 import com.google.firebase.firestore.FirebaseFirestore
-import javax.inject.Inject
 import kotlinx.coroutines.tasks.await
+import javax.inject.Inject
 
 class MarketRemoteDataSource @Inject constructor(
     private val firestore: FirebaseFirestore
