+++ app/src/main/java/com/delighted2wins/souqelkhorda/features/market/presentation/component/orderdetails/OrderItemCard.kt
@@ -1,7 +1,13 @@
 package com.delighted2wins.souqelkhorda.features.market.presentation.component.orderdetails
 
 import android.annotation.SuppressLint
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Card
 import androidx.compose.material3.CardDefaults
